// src/app/interfaces/crime.ts
export interface ICrime {
  key?: string;
  crime_no?: string;
  crime_commune_no?: number;
  crime_year_no?: number;
  crime_index?: number;
  category?: any;
  sub_category?: any;
  case_happen_date: Date;
  case_happen_date_key: number;
  case_happen_date_time: string;
  case_happen_date_time_number?: number;
  case_receive_date: Date;
  case_receive_date_key: number;
  case_receive_date_time: string;
  case_receive_date_time_number?: number;
  case_receive_from?: string;
  province?:any;
  district?:any;
  commune?:any;
  village?:any;
  map: any;
  road:string;
  house_no:string;
  location:any;
  location_scene: any;
  location_name: string;
  case_reason: any;
  case_reason_other: string;
  weapon: any;
  weapon_other: any;
  transportation: any;
  transportation_other: any;
  victim_total: number;
  victim_female_total: number;
  suspect_total: number;
  suspect_female_total: number;
  suspect_arrested_total: number;
  suspect_arrested_female_total: number;
  notifications?: any;
  tags?: any;
  status?: any;
  crime_status?: any;
  page_key?: number;
  create_date?: Date;
  create_date_key?: number;
  create_date_month?: string;
  create_date_year?: string;
  create_by?: any;
  update_date?: Date;
  update_by?: any;
}

export interface ICrimeTransfer {
  key?: string;
  status?: any;
  page_key?: number;
  create_date?: Date;
  create_by?: any;
  province?: any;
  district?: any;
  commune?: any;
  village?: any;
  owner_by?: any;
  transfer_to?: any;
  transfer_by?: any;
  transfer_date?: Date;
  transfer_date_key?: number;
  note?: string;
}